pyproject.toml
[project]
name = "gateway-privacidade"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dotenv>=0.9.9",
    "numpy>=2.0.2",
    "paho-mqtt>=2.1.0",
    "redis>=6.4.0",
    "requests>=2.32.5",
]
